Test Cleanup: Improve test setup
Use `touch` test helper instead of directly working with the `File`
class during the setup phase.

Also create [file_fixtures][] to simplify test setup.

Because the newly added file fixtures are processed with standard, I
needed to adjust the testing generator to account for a minor violation.

[file_fixtures]: https://api.rubyonrails.org/classes/ActiveSupport/Testing/FileFixtures.html#method-i-file_fixture
