+++ test/fixtures/files/Rakefile
@@ -1,0 +1,5 @@
+require_relative "config/application"
+require "bundler/audit/task"
+Bundler::Audit::Task.new
+
+Rails.application.load_tasks
