+++ test/fixtures/files/_flashes.html.erb
@@ -1,0 +1,7 @@
+<% if flash.any? %>
+  <div class="flashes">
+    <% flash.each do |type, message| -%>
+      <div class="flash-<%= type %>"><%= message %></div>
+    <% end -%>
+  </div>
+<% end %>
