+++ test/fixtures/files/action_mailer.rb
@@ -1,0 +1,5 @@
+RSpec.configure do |config|
+  config.before(:each) do
+    ActionMailer::Base.deliveries.clear
+  end
+end
