+++ test/fixtures/files/email_interceptor_initializer.rb
@@ -1,0 +1,5 @@
+Rails.application.configure do
+  if ENV["INTERCEPTOR_ADDRESSES"].present?
+    config.action_mailer.interceptors = %w[EmailInterceptor]
+  end
+end
