+++ test/fixtures/files/factories.rb
@@ -1,0 +1,2 @@
+FactoryBot.define do
+end
