+++ test/fixtures/files/factories_spec_lint.rb
@@ -1,0 +1,7 @@
+require "rails_helper"
+
+RSpec.describe "Factories" do
+  it "has valid factoties" do
+    FactoryBot.lint traits: true
+  end
+end
