+++ test/fixtures/files/factories_test_lint.rb
@@ -1,0 +1,9 @@
+require "test_helper"
+
+class FactoryBotsTest < ActiveSupport::TestCase
+  class FactoryLintingTest < FactoryBotsTest
+    test "linting of factories" do
+      FactoryBot.lint traits: true
+    end
+  end
+end
