+++ test/fixtures/files/i18n.rb
@@ -1,0 +1,3 @@
+RSpec.configure do |config|
+  config.include ActionView::Helpers::TranslationHelper
+end
