+++ test/fixtures/files/inline_svg.rb
@@ -1,0 +1,3 @@
+InlineSvg.configure do |config|
+  config.raise_on_file_not_found = true
+end
