+++ test/fixtures/files/prettierignore
@@ -1,0 +1,1 @@
+vendor/bundle/**
