+++ test/fixtures/files/stylelintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@thoughtbot/stylelint-config"
+}
